:bug: Fix OCI client configuration logic
This commit fixes the logic deciding to create a TLS client for OCI
charts.

The existing code was creating a non-TLS configured client if either the CA file
was unspecified or if `insecureSkipTLSVerify` was set to false. If a CA
file was specified then `insecureSkipTLSVerify` is false, which meant
that a non-TLS client was always created, causing cert validation
failures if the OCI registry is served over TLS.

This commit changes the logic to create a non-TLS configured client if
both CA file is unset and `insecureSkipTLSVerify` is false.